healthcare_backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
awsebcli==3.21.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
botocore==1.35.99
cachetools==5.5.2
cement==2.10.14
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.3.2
cryptography==44.0.1
cycler==0.12.1
daphne==4.1.2
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Django==5.0.7
django-ckeditor==6.7.2
django-cors-headers==4.4.0
django-environ==0.12.0
django-js-asset==3.0.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
fonttools==4.58.0
fpdf==1.7.2
fqdn==1.5.1
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.68.0
googletrans==4.0.2
greenlet==3.2.1
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mistune==3.1.3
mysqlclient==2.2.6
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.2
notebook_shim==0.2.4
numpy==2.2.5
oauthlib==3.2.2
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.10.1
pillow==10.4.0
platformdirs==4.3.8
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
pypiwin32==223
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-json-logger==3.3.0
pytz==2025.1
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
rsa==4.9
seaborn==0.13.2
selenium==4.31.0
semantic-version==2.10.0
Send2Trash==1.8.3
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.40
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.46.2
termcolor==2.5.0
terminado==0.18.1
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
Twisted==24.11.0
txaio==23.1.1
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.34.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
whitenoise==6.7.0
wsproto==1.2.0
zope.interface==7.2
